Renamed all the files from "discover*" to "app_discover* to comply with review comments
