+++ modules/jaggery-apps/app-publisher-web/jaggery.conf
@@ -14,7 +14,7 @@
 		},
 		{
 			"url":"/assets/discover/*",
-			"path":"/controllers/assets_discover_router.jag"
+			"path":"/controllers/app_discover_router.jag"
 		},
 		{
 			"url":"/api/asset/*",
@@ -26,7 +26,7 @@
         },
         {
 			"url":"/api/discover/*",
-			"path":"/apis/v1/discover_api_router.jag"
+			"path":"/apis/v1/app_discover_api_router.jag"
 		},
 		{
 			"url":"/api/lifecycle/*",
