+++ modules/jaggery-apps/app-publisher-web/themes/appm/helpers/app_discover.js
@@ -1,6 +1,6 @@
 var resources=function(page,meta){
     return{
-        js:['caramel.handlebars.client.js', '/logic/recent-assets.js', 'discover.js'],
+        js:['caramel.handlebars.client.js', '/logic/recent-assets.js', 'app_discover.js'],
         css: ['webapp/styles.css']
