+++ modules/jaggery-apps/app-publisher-web/themes/appm/partials/app_discover-status.hbs
