+++ modules/jaggery-apps/app-publisher-web/themes/appm/partials/app_discover.hbs
@@ -142,5 +142,5 @@
     {{/if}}
 
 </div>
-{{> discover-create-status}}
+{{> app_discover-status}}
 </div>
